fix: on ne supprime plus un utilisateur lorsqu'on supprime sa relation à un ETG en tant qu'admin (#450)
* fix: on arrête de checker window.navigator.onLine on utilise le store

* fix: on ne supprime plus un utilisateur lorsqu'on supprime sa relation à un ETG

+++ api-express/src/controllers/user-entity.ts
@@ -403,12 +403,14 @@ router.delete(
           },
         });
         if (user && !user.last_login_at) {
-          // user was wrongly invited, we need to delete him
-          await prisma.user.delete({
-            where: {
-              id: body.owner_id,
-            },
-          });
+          if (!req.user.isZacharieAdmin) {
+            // user was wrongly invited, we need to delete him
+            await prisma.user.delete({
+              where: {
+                id: body.owner_id,
+              },
+            });
+          }
         }
       }
